+++ app/src/main/assets/CHANGELOG.md
@@ -8,7 +8,8 @@ For support you can use [the contact form](https://contact.faircode.eu/?product=
 
 ### Next version
 
-* Added mini audio player (display-settings tab page)
+* Added mini audio player
+* Added video preview image
 * Small improvements and minor bug fixes
 * Updated [AndroidX](https://developer.android.com/jetpack/androidx/versions/all-channel)
 * Updated [Public Suffix List](https://github.com/publicsuffix/list)